Replace magit autopush/pull with bash script
It still wasn't working. If there were changes, they would only commit
and not push. I had to run it twice to push the changes, which is
dumb. Plus magit is meant to be front-end thing, was never meant to be
run programatically. Better to use a bash script here.